% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_counts}
\alias{add_counts}
\title{Add counts to the levels of a factor}
\usage{
add_counts(f)
}
\arguments{
\item{f}{a factor variable}
}
\description{
Add counts to the levels of a factor
}

