% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2.R
\name{construct_R2s}
\alias{construct_R2s}
\title{construct R2 across all conditions}
\usage{
construct_R2s(m, intensity)
}
\arguments{
\item{m}{normalized matrix}

\item{intensity}{vector of intensities}
}
\description{
construct R2 across all conditions
}

