% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pseudoR2.R
\name{constructPseudoR2}
\alias{constructPseudoR2}
\title{calculate pseudo-R^2 for every row based on a dispersion model}
\usage{
constructPseudoR2(mat, disp)
}
\arguments{
\item{mat}{count matrix}

\item{disp}{Calculated per-gene dispersion, or 0 for a Poisson model}
}
\description{
calculate pseudo-R^2 for every row based on a dispersion model
}
\details{
This function is largely from the pre-refactoring version and could
use some further reorganization.
}

