% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{do_row}
\alias{do_row}
\title{perform an operation on each line of a rowwise data frame}
\usage{
do_row(x, ..., groups = NULL)
}
\arguments{
\item{x}{a tbl}

\item{...}{Additional arguments passed on to \code{do}}

\item{groups}{Vector of columns to group by. If missing,
will attempt to guess based on all columns that are not lists.}
}
\description{
A \code{\link{do}} operation given multiple arguments creates a rowwise
data frame. Often one wants to follow this with an operation on each line
of the data frame, but this requires regrouping (which leads to an error),
then performing an awkward expression with lots of \code{.$column[[1]]}
in it. This is a simple wrapper of \code{do} that makes that easier.
}
\examples{
library(dplyr)
library(broom)

g <- mtcars \%>\% group_by(cyl) \%>\% do(m = lm(mpg ~ wt, .))
g \%>\% do_row(tidy(m))
g \%>\% do_row(augment(m))
g \%>\% do_row(tidied = tidy(m), augmented = augment(m))
}

