% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_sided_violin.R
\name{geom_sided_violin}
\alias{geom_sided_violin}
\title{a violin plot with two sides}
\usage{
geom_sided_violin(mapping = NULL, data = NULL, stat = "ydensity",
  position = "dodge", trim = TRUE, scale = "area", ...)
}
\description{
A sided violin plot is good for comparing two distributions across
many sets, such as comparing microarray and RNA-Seq for each set
}

