% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeMatrix.matrix}
\alias{normalizeMatrix.matrix}
\title{Normalize an RNA-Seq matrix}
\usage{
\method{normalizeMatrix}{matrix}(mat, normalize.method = "scale")
}
\description{
Normalize an RNA-Seq matrix
}

