% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{Normalize an RS/MA matrix according to the appropriate normalization method}
\usage{
normalizeMatrix(mat, normalize.method = "scale")
}
\description{
Normalize an RS/MA matrix according to the appropriate normalization method
}

