% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perform_edgeR.R
\name{perform_edgeR}
\alias{perform_edgeR}
\title{Perform edgeR exact test on a matrix, given the groups}
\usage{
perform_edgeR(matrix, group = c("E", "E", "G", "G"))
}
\description{
Perform edgeR exact test on a matrix, given the groups
}

