% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pool_matrix.R
\name{pool_matrix.matrix}
\alias{pool_matrix.matrix}
\title{Given a microarray/RNASeq matrix and a subset of columns, pool within columns}
\usage{
\method{pool_matrix}{matrix}(m, des, cond, average = FALSE)
}
\arguments{
\item{m}{a matrix of RNA-Seq counts or microarray logFCs}

\item{des}{Experiment design (such as built-in experimentDesign)}

\item{cond}{vector used to subset expression matrix and design}

\item{average}{Whether values should be averaged: otherwise they are summed}
}
\value{
The matrix, using the given subset, pooled within condition:bio
}
\description{
Given a microarray/RNASeq matrix and a subset of columns, pool within columns
}

