% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pool_matrix.R
\name{pool_matrix.RGList}
\alias{pool_matrix.RGList}
\title{method for pooling an RGList object}
\usage{
\method{pool_matrix}{RGList}(m, des, cond, average = TRUE)
}
\description{
method for pooling an RGList object
}

