% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2.R
\name{R2s}
\alias{R2s}
\title{compute R-squared values for many genes}
\usage{
R2s(m, mm.full, adjusted = TRUE)
}
\arguments{
\item{m}{a normalized matrix from normalizeBetweenArrays or voom}

\item{mm.full}{the model matrix whose R^2 is being examined}

\item{adjusted}{whether the R^2 values should be adjusted for degrees of freedom}
}
\description{
compute R-squared values for many genes
}

